Move entity context provider into context directory
