+++ src/app/Authenticated.tsx
@@ -9,8 +9,8 @@ import CaptureEdit from 'components/capture/Edit';
 import FullPageSpinner from 'components/fullPage/Spinner';
 import MaterializationCreate from 'components/materialization/Create';
 import MaterializationEdit from 'components/materialization/Edit';
-import { EntityContextProvider } from 'components/shared/Entity/EntityContext';
 import AuthenticatedOnlyContext from 'context/Authenticated';
+import { EntityContextProvider } from 'context/EntityContext';
 import { OAuthPopup } from 'hooks/forks/react-use-oauth2/components';
 import useCombinedGrantsExt from 'hooks/useCombinedGrantsExt';
 import useGatewayAuthToken from 'hooks/useGatewayAuthToken';
